+++ src/components/WhiteNoisePlayer.tsx
@@ -102,7 +102,7 @@ class NoiseGenerator {
     const buffer = this.audioContext.createBuffer(1, 44100 * 2, 44100)
     const noiseData = this.generateNoise(type)
 
-    buffer.copyToChannel(noiseData, 0)
+    buffer.copyToChannel(noiseData as Float32Array, 0)
 
     // Crear nodos
     this.source = this.audioContext.createBufferSource()
